Create widgets folder and move dotWidget that folder
